requirements.txt
Flask
requests
beautifulsoup4
googlesearch-python
python-dotenv
